Move helper functions & definitions from search-header to helper.ts
To clean up the component a bit.
