+++ src/components/autosuggest/autosuggest.tsx
@@ -6,7 +6,6 @@ import { Suggestion, Suggestions } from "../../core/utils/types/autosuggest";
 import AutosuggestCategory from "../autosuggest-category/autosuggest-category";
 import AutosuggestMaterial from "../autosuggest-material/autosuggest-material";
 import { AutosuggestText } from "../autosuggest-text/autosuggest-text";
-import { AutosuggestCategoryList } from "../../core/utils/types/material-type";
 
 export interface AutosuggestProps {
   textData: SuggestionsFromQueryStringQuery["suggest"]["result"];
@@ -17,7 +16,6 @@ export interface AutosuggestProps {
   getItemProps: UseComboboxPropGetters<Suggestion>["getItemProps"];
   isOpen: boolean;
   categoryData?: SuggestionsFromQueryStringQuery["suggest"]["result"];
-  autosuggestCategoryList: AutosuggestCategoryList[];
   isLoading: boolean;
   dataCy?: string;
 }
@@ -30,7 +28,6 @@ export const Autosuggest: React.FC<AutosuggestProps> = ({
   getItemProps,
   isOpen,
   categoryData,
-  autosuggestCategoryList,
   isLoading,
   dataCy = "autosuggest"
 }) => {
@@ -75,7 +72,6 @@ export const Autosuggest: React.FC<AutosuggestProps> = ({
             getItemProps={getItemProps}
             highlightedIndex={highlightedIndex}
             textAndMaterialDataLength={textData.length + materialData.length}
-            autosuggestCategoryList={autosuggestCategoryList}
           />
         )}
       </ul>
